+++ src/node/src/node.rs
@@ -16,6 +16,7 @@ use async_dup::Arc as AsyncArc;
 use async_io::Async;
 use protocol::{Command, ProtocolConfiguration, Receiver};
 use smol::{self, block_on, Task};
+use sql_engine::frontend::FrontendStorage;
 use sql_engine::QueryExecutor;
 use std::{
     env,
@@ -26,7 +27,7 @@ use std::{
         Arc, Mutex,
     },
 };
-use storage::{backend::SledBackendStorage, frontend::FrontendStorage};
+use storage::SledBackendStorage;
 
 const PORT: u16 = 5432;
 const HOST: [u8; 4] = [0, 0, 0, 0];
