+++ src/sql_engine/Cargo.toml
@@ -15,6 +15,8 @@ sql_types = { path = "../sql_types" }
 protocol = { path = "../protocol" }
 futures-lite = "0.1.11"
 representation = { path = "../representation" }
+serde = { version = "1.0.114", features = ["derive"] }
+bincode = "1.3.1"
 
 [dev-dependencies]
 rstest = "0.6.4"
