+++ src/sql_engine/src/lib.rs
@@ -15,6 +15,7 @@
 extern crate bigdecimal;
 extern crate log;
 
+use crate::frontend::FrontendStorage;
 use crate::{
     ddl::{
         create_schema::CreateSchemaCommand, create_table::CreateTableCommand, drop_schema::DropSchemaCommand,
@@ -28,13 +29,97 @@ use protocol::{
     results::{QueryErrorBuilder, QueryEvent},
     Sender,
 };
+use serde::{Deserialize, Serialize};
+use sql_types::SqlType;
 use sqlparser::{ast::Statement, dialect::PostgreSqlDialect, parser::Parser};
 use std::sync::{Arc, Mutex};
-use storage::{backend::BackendStorage, frontend::FrontendStorage};
+use storage::BackendStorage;
 
 mod ddl;
 mod dml;
+pub mod frontend;
 mod query;
+
+pub type Projection = (Vec<ColumnDefinition>, Vec<Vec<String>>);
+
+#[derive(Debug, Clone)]
+pub struct TableDescription {
+    schema_name: String,
+    table_name: String,
+    column_data: Vec<ColumnDefinition>,
+}
+
+impl TableDescription {
+    pub fn new(schema_name: &str, table_name: &str, column_data: Vec<ColumnDefinition>) -> Self {
+        Self {
+            schema_name: schema_name.to_owned(),
+            table_name: table_name.to_owned(),
+            column_data,
+        }
+    }
+
+    pub fn column_len(&self) -> usize {
+        self.column_data.len()
+    }
+
+    pub fn column_type(&self, column_idx: usize) -> SqlType {
+        if let Some(column) = self.column_data.get(column_idx) {
+            column.sql_type
+        } else {
+            panic!("attempting to access type of invalid column index")
+        }
+    }
+
+    pub fn column_type_by_name(&self, name: &str) -> Option<SqlType> {
+        self.column_data
+            .iter()
+            .find(|column| column.name == name)
+            .map(|column| column.sql_type)
+    }
+
+    pub fn column_data(&self) -> &[ColumnDefinition] {
+        self.column_data.as_slice()
+    }
+
+    pub fn scheme(&self) -> &str {
+        self.schema_name.as_str()
+    }
+
+    pub fn table(&self) -> &str {
+        self.table_name.as_str()
+    }
+
+    pub fn full_name(&self) -> String {
+        format!("{}.{}", self.schema_name, self.table_name)
+    }
+}
+
+#[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
+pub struct ColumnDefinition {
+    name: String,
+    sql_type: SqlType,
+}
+
+impl ColumnDefinition {
+    pub fn new(name: &str, sql_type: SqlType) -> Self {
+        Self {
+            name: name.to_string(),
+            sql_type,
+        }
+    }
+
+    pub fn sql_type(&self) -> SqlType {
+        self.sql_type
+    }
+
+    pub fn has_name(&self, other_name: &str) -> bool {
+        self.name == other_name
+    }
+
+    pub fn name(&self) -> String {
+        self.name.clone()
+    }
+}
 
 pub struct QueryExecutor<P: BackendStorage> {
     storage: Arc<Mutex<FrontendStorage<P>>>,
