+++ src/sql_engine/src/query/process.rs
@@ -12,9 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use crate::frontend::FrontendStorage;
 ///! Module for transforming the input Query AST into representation the engine can process.
 use crate::query::plan::{Plan, SchemaCreationInfo, TableCreationInfo, TableInserts};
 use crate::query::{SchemaId, SchemaNamingError, TableId, TableNamingError};
+use crate::ColumnDefinition;
 use protocol::{results::QueryErrorBuilder, Sender};
 use sql_types::SqlType;
 use sqlparser::ast::{ColumnDef, DataType, ObjectName, ObjectType, Statement};
@@ -22,7 +24,7 @@ use std::{
     convert::TryFrom,
     sync::{Arc, Mutex, MutexGuard},
 };
-use storage::{backend::BackendStorage, frontend::FrontendStorage, ColumnDefinition};
+use storage::BackendStorage;
 
 type Result<T> = std::result::Result<T, ()>;
